+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- feat: provide a standalone PHAR distribution for CI and non-Composer projects
+
 ## [1.5.0] - 2026-06-29
 
 - feat: add target-language consistency check for XLIFF files (filename locale vs. declared target-language)